responsive design des pages utilisateurs non authentifié
